+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Add [TextOverflowWidget.position] to support show overflow at start, middle or end.
   https://github.com/flutter/flutter/issues/45336
-* Add [ExtendedText.betterLineBreakingAndOverflowStyle] to make line breaking and overflow style better.
+* Add [ExtendedText.joinZeroWidthSpace] to make line breaking and overflow style better.
   https://github.com/flutter/flutter/issues/18761
 * Fix strutStyle not work.
 * Breaking change: remove [TextOverflowWidget.fixedOffset]
@@ -13,7 +13,27 @@
 * fix strutStyle miss
 ## 5.0.0-non-null-safety
 
-* non-null-safety
+## 5.0.4
+
+* Fix issue that the overflowWidget is not layout #97
+
+## 5.0.3
+
+* Fix null-safety error #96
+
+## 5.0.2
+
+* Fix null-safety error
+
+## 5.0.1
+
+* Add add SemanticsInformation for overflowWidget
+* Improve performance for overflowWidget
+* Do not paint Selection in the region of overFlowWidget
+
+## 5.0.0
+
+* Support null-safety
 
 ## 4.1.0